pyproject.toml
[project]
name = "frontier-cs"
version = "0.1.0"
description = "Evaluation framework for Frontier-CS problems"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.74.0",
    "colorlog>=6.10.1",
    "datasets>=4.4.1",
    "google-genai>=1.55.0",
    "google-generativeai>=0.8.5",
    "numpy==2.3.4",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "requests>=2.32.5",
    "skypilot>=0.10.5",
    "tqdm>=4.67.1",
]

[project.scripts]
frontier = "frontier_cs.cli:main"
frontier-eval = "frontier_cs.cli:main"  # deprecated alias

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
